Fixes #72 Refactor module structure to layered architecture
