+++ src/agent/entity_ai.rs
