+++ src/game.rs
@@ -2,9 +2,9 @@ pub mod component;
 pub mod config;
 pub mod engagement;
 pub mod entity;
-pub mod entity_ai;
 pub mod game_loop;
 pub mod game_state;
+pub mod interaction;
 pub mod mailbox;
 pub mod map;
 pub mod messaging;
@@ -47,9 +47,6 @@ pub use engagement::TurnAction;
 pub use engagement::TurnOrder;
 pub use entity::Entity;
 pub use entity::EntityType;
-pub use entity_ai::ConversationContext;
-pub use entity_ai::EntityAI;
-pub use entity_ai::SimpleConversationState;
 pub use game_state::GameState;
 pub use mailbox::Mailboxes;
 pub use map::Dungeon;
