+++ src/game/engagement/conversation/agent_turn.rs
@@ -1,8 +1,8 @@
 use std::sync::Arc;
 
 use crate::agent;
+use crate::agent::entity_ai::{AgentMessage, AgentRole};
 use crate::agent::tools::InspectEntity;
-use crate::game::entity_ai::{AgentMessage, AgentRole};
 use crate::game::player::Player;
 use crate::game::{GameState, messaging};
