+++ src/network.rs
@@ -2,5 +2,6 @@ pub mod client;
 pub mod discovery;
 pub mod event;
 pub mod server;
+pub mod session;
 
 pub use event::NetworkEvent;
