+++ src/network/server/state.rs
@@ -7,8 +7,8 @@ use std::time::Instant;
 
 use crate::game::GameState;
 use crate::network::event::NetworkEvent;
+use crate::network::session::ServerSession;
 use crate::persistence::Database;
-use crate::session::ServerSession;
 use futures_util::stream::Stream;
 use serde::Deserialize;
 use tokio::sync::RwLock;
