+++ src/network/session/client_session.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 use super::error::SessionError;
-use crate::state::config;
+use crate::paths;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ClientSession {
@@ -11,7 +11,7 @@ pub struct ClientSession {
 
 impl ClientSession {
     pub async fn load(server_id: &str) -> Result<Option<Self>, SessionError> {
-        let path = config::client_session_file(server_id).map_err(|_| SessionError::NoHomeDir)?;
+        let path = paths::client_session_file(server_id).map_err(|_| SessionError::NoHomeDir)?;
         if !path.exists() {
             return Ok(None);
         }
@@ -21,7 +21,7 @@ impl ClientSession {
     }
 
     pub async fn save(&self, server_id: &str) -> Result<(), SessionError> {
-        let path = config::client_session_file(server_id).map_err(|_| SessionError::NoHomeDir)?;
+        let path = paths::client_session_file(server_id).map_err(|_| SessionError::NoHomeDir)?;
         let data = serde_json::to_string_pretty(self)?;
         tokio::fs::write(&path, data).await?;
         Ok(())
