+++ src/network/session/error.rs
