+++ src/paths.rs
@@ -50,12 +50,10 @@ pub async fn create_session_base_dirs() -> StateResult<()> {
 pub fn find_config_dir() -> Option<PathBuf> {
     let cwd = std::env::current_dir().ok()?;
 
-    // Check working directory itself
     if cwd.join("mud.toml").exists() {
         return Some(cwd);
     }
 
-    // Check immediate subdirectories of muds/
     let muds_dir = cwd.join("muds");
     if muds_dir.is_dir()
         && let Ok(entries) = std::fs::read_dir(&muds_dir)
