--- src/state.rs
@@ -1,6 +1,0 @@
-pub mod config;
-
-pub use config::{
-    client_session_dir, client_session_file, create_session_base_dirs, create_state_dirs,
-    mudroom_dir, server_session_dir, server_session_file, session_dir,
-};
