+++ src/tui/screens/discovery.rs
